removed repository pattern
