--- ThemingCodelab/app/src/main/java/com/example/reply/data/EmailsRepositoryImpl.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.reply.data
-
-import com.example.reply.data.local.LocalEmailsDataProvider
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-
-class EmailsRepositoryImpl : EmailsRepository {
-
-    override fun getAllEmails(): Flow<List<Email>> = flow {
-        emit(LocalEmailsDataProvider.allEmails)
-    }
-
-    override fun getCategoryEmails(category: MailboxType): Flow<List<Email>> = flow {
-        val categoryEmails = LocalEmailsDataProvider.allEmails.filter { it.mailbox == category }
-        emit(categoryEmails)
-    }
-
-    override fun getAllFolders(): List<String> {
-        return LocalEmailsDataProvider.getAllFolders()
-    }
-
-    override fun getEmailFromId(id: Long): Flow<Email?> = flow {
-        val categoryEmails = LocalEmailsDataProvider.allEmails.firstOrNull { it.id == id }
-    }
-}
