+++ ThemingCodelab/app/src/main/java/com/example/reply/ui/MainActivity.kt
@@ -26,13 +26,12 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.reply.data.local.LocalEmailsDataProvider
+import com.example.reply.data.LocalEmailsDataProvider
 import com.example.reply.ui.theme.AppTheme
 
 class MainActivity : ComponentActivity() {
 
     private val viewModel: ReplyHomeViewModel by viewModels()
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
